requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
appdirs=1.4.4=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
autograd=1.6.2=pypi_0
autoray=0.6.12=pypi_0
blas=1.0=mkl
brotli=1.0.9=h5eee18b_8
brotli-bin=1.0.9=h5eee18b_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.9.24=h06a4308_0
cachetools=5.5.0=pypi_0
certifi=2024.7.4=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
comm=0.2.1=py311h06a4308_0
contourpy=1.2.0=py311hdb19cb5_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.6.37=0
cuda-runtime=12.1.0=0
cuda-version=12.6=3
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
cytoolz=1.0.0=pypi_0
data=0.4=pypi_0
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=py311h6a678d5_0
decorator=5.1.1=pyhd3eb1b0_0
diastatic-malt=2.15.2=pypi_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.6.2=h6a678d5_0
filelock=3.13.1=py311h06a4308_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.51.0=py311h5eee18b_0
freetype=2.12.1=h4a9f257_0
funcsigs=1.0.2=pypi_0
future=1.0.0=pypi_0
gast=0.6.0=pypi_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=73.1=h6a678d5_0
idna=3.7=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.28.0=py311h06a4308_0
ipython=8.25.0=py311h06a4308_0
jax=0.4.23=pypi_0
jaxlib=0.4.23=pypi_0
jedi=0.19.1=py311h06a4308_0
jinja2=3.1.4=py311h06a4308_0
jpeg=9e=h5eee18b_3
jupyter_client=8.6.0=py311h06a4308_0
jupyter_core=5.7.2=py311h06a4308_0
kiwisolver=1.4.4=py311h6a678d5_0
krb5=1.20.1=h143b758_1
latex=0.7.0=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.11.0.15=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.7.37=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libllvm14=14.0.6=hdb19cb5_3
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h097e994_2
libxml2=2.13.1=hfdd30dd_2
llvm-openmp=14.0.6=h9e868ea_0
llvmlite=0.43.0=py311h6a678d5_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py311h5eee18b_0
matplotlib=3.8.4=py311h06a4308_0
matplotlib-base=3.8.4=py311ha02d727_0
matplotlib-inline=0.1.6=py311h06a4308_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
ml-dtypes=0.4.0=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
mysql=5.7.24=h721c034_2
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=py311h06a4308_0
networkx=3.3=py311h06a4308_0
numba=0.60.0=py311ha02d727_0
numpy=1.26.4=py311h08b1b3b_0
numpy-base=1.26.4=py311hf175353_0
opencv-python-headless=4.10.0.84=pypi_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.0.15=h5eee18b_0
opt-einsum=3.3.0=pypi_0
packaging=24.1=py311h06a4308_0
parso=0.8.3=pyhd3eb1b0_0
pcre2=10.42=hebb0a14_1
pennylane=0.37.0=pypi_0
pennylane-catalyst=0.7.0=pypi_0
pennylane-lightning=0.37.0=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pillow=10.4.0=py311h5eee18b_0
pip=24.2=py311h06a4308_0
platformdirs=3.10.0=py311h06a4308_0
ply=3.11=py311h06a4308_0
prompt-toolkit=3.0.43=py311h06a4308_0
prompt_toolkit=3.0.43=hd3eb1b0_0
psutil=5.9.0=py311h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pybind11-abi=4=hd3eb1b0_1
pygments=2.15.1=py311h06a4308_1
pyparsing=3.0.9=py311h06a4308_0
pyqt=5.15.10=py311h6a678d5_0
pyqt5-sip=12.13.0=py311h5eee18b_0
python=3.11.9=h955ad1f_0
python-dateutil=2.9.0post0=py311h06a4308_2
pytorch=2.4.0=py3.11_cuda12.1_cudnn9.1.0_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pyyaml=6.0.1=py311h5eee18b_0
pyzmq=25.1.2=py311h6a678d5_0
qt-main=5.15.2=h53bd1ea_10
readline=8.2=h5eee18b_0
requests=2.32.3=pypi_0
ruamel-yaml=0.17.40=pypi_0
ruamel-yaml-clib=0.2.12=pypi_0
rustworkx=0.15.1=pypi_0
scipy=1.12.0=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=72.1.0=py311h06a4308_0
shutilwhich=1.1.0=pypi_0
sip=6.7.12=py311h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
symfit=0.5.6=pypi_0
sympy=1.12=py311h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tempdir=0.7.1=pypi_0
termcolor=2.4.0=pypi_0
tk=8.6.14=h39e8969_0
toml=0.10.2=pypi_0
toolz=1.0.0=pypi_0
toposort=1.10=pypi_0
torchtriton=3.0.0=py311
tornado=6.4.1=py311h5eee18b_0
traitlets=5.14.3=py311h06a4308_0
typing_extensions=4.11.0=py311h06a4308_0
tzdata=2024a=h04d1e81_0
unicodedata2=15.1.0=py311h5eee18b_0
urllib3=2.2.2=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.43.0=py311h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zeromq=4.3.5=h6a678d5_0
zlib=1.2.13=h5eee18b_1
zstandard=0.23.0=pypi_0
zstd=1.5.5.1=pypi_0
